Added ordering to Scale, added virtual smobs, Removed pitch as smobtype.
Added information to scale about how pitchclasses with the same accidental
should be ordered. (That is, in the case of default scale, F-C-G-D-A-E-B).
Using this, the old code to ordering accidentals in the key signature could
be discarded, afaics. Have not done this yet.

Added virtual smobs, thereby supporting polymorphic smobs.

Removed pitch as smobtpype leaving Pitchclass as the smobtype of pitches
and pitchclasses.