Fix Issue 2529 "configure should error for any missing software"
This is not an optimal fix, as by the moment this change is committed,
GUB bypasses configure for enabling documentation build, so
developers/packagers who call "configure --disable-documentation" will
still get warnings about missing requirements for documentation
build. The issues this patch addresses are
* adding missing documentation build requirements,
* enforcing all documentation build requirements in case
documentation build is enabled in configure.