Let #{ 4 #} be a duration while #{ 4 \cm #} stays a length
An UNSIGNED not followed by a NUMERIC_IDENTIFIER inside of #{ ... #}
is interpreted as a duration rather than a number.  Also, inside
of #{ ... #} no "numeric expressions" using binary operators +-*/ are
interpreted.

The behavior of (signed and unsigned) integers and numeric expressions
in assignments and as function arguments rather than inside
of #{ ... #} is not affected.