font: clean up staffline-display in testing mode
Previous testing code was a mix of hideous copy&paste with
non-obvious passing glyph outlines around, which attempted
to produce additional glyphs in testing mode, so that there
would be both on-staffline and on-staffspace variations.

That code was completely unreadable and unmaintainable.

Instead i've introduced a global variable that determines
how the stafflines will be printed relative to the glyphs.
To see alternative configuration, just change that value.