add \setTextFont
This function has to be used to set text fonts when
\setNotationFont has been used with the "text-fonts = none"
option (use case: stylesheets that don't impose specific fonts).
In this case it is necessary to give the user a chance to
set text fonts afterwards.

TODO: I want to check if add-pango-fonts can be modified
(or a function added) that adds or replaces a font node.
This would make resetting the font tree obsolete, as well
as the text-fonts = none approach (because then any text
font could simply be *changed* instead of being added on
top of an empty tree).