Rename DATADIR to CONFIG_DATADIR
The header objidl.h from mingw has an enum called DATADIR. This
conflicts with our definition and results in compilation errors
because the preprocessor replaces the enum name with the configured
string. This is apparently dealt with in mingw-compatibility.hh,
but not in the configure checks. As a result, they return the wrong
answer which leads to further compilation errors down the road.